handling rehydrate
