+++ src/aind_data_schema/imaging/acquisition.py
@@ -3,10 +3,9 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 from typing import List, Optional
-
-from decimal import Decimal
 
 from pydantic import Field
